started refactoring the MapActivity with the MapFragment. I'm not getting any errors, but when I run it, the list will just reload the list and will not take me to the map when a tast is clicked
